o Added new snapping to pictures o Fix painting during insert of rect, ellipse, pie when scrolled
svn path=/trunk/koffice/; revision=478234

+++ kpresenter/KPresenterPageIface.h
@@ -103,7 +103,7 @@ k_dcop:
     DCOPRef insertLineD1( int x,int y, int h, int w, bool rev );
     DCOPRef insertLineD2( int x,int y, int h, int w, bool rev );
     DCOPRef insertTextObject( int x,int y, int h, int w );
-    DCOPRef insertPicture( const QString & file,int x,int y, int h, int w );
+    DCOPRef insertPicture( const QString & file, double x, double y, double h, double w );
 
     void deSelectAllObj();
